# JIT compiler, first cut