… src/methods.ts
… src/opcodes.ts
… src/threading.ts