@hrj 2016-04-21
The source code and binaries used for benchmarking, along with a README, are available [here](https://hrj.github.io/static/benchmark-files.tar.bz2).