@perimosocordiae 2016-04-26
Impressive work! It's a lot to review, so one thing that would help a lot is if you did some rebasing to structure the commits a little more nicely. Ideally, there should be:
 1. a commit that makes the non-JIT changes (like the new convenience functions for stack manipulation, and the signature change for the opcode annotation)
 2. a commit with the JIT, in a minimal working state
 3. commit(s) adding implementations of the various opcodes

It's a bit of a pain to shuffle commits around, I know, but it really does make a giant PR like this one easier to understand.