@hrj 2016-04-26
> It's a bit of a pain to shuffle commits around, I know, but it really does make a giant PR like this one easier to understand.

I agree, and thanks for looking!

@jvilk I am glad you mentioned a date.

I will work on a more structured PR and try to push it here before May 10th.