@hrj 2016-05-06
superseded by #443